requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0

# API & Async
httpx==0.25.2
asyncio==3.4.3
python-multipart==0.0.6
openai>=1.40.0

# Image Processing
pillow==10.1.0

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Data handling
pandas==2.1.3
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
alembic==1.12.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0